requirements.txt
telethon>=1.40.0
openpyxl>=3.1.0
sqlalchemy>=2.0.0
python-dotenv>=0.19.0
asyncio
